test: moving test folder out of the src folder to be able to run the project with missing controllers used on tests
