+++ test/controllers/auth/login.test.ts
@@ -1,6 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AuthController } from '@/infrastructure/http/controllers/auth.controller';
+import { AuthController } from '@/infrastructure/http/controllers/auth/auth.controller';
 import { LoginUseCase } from '@/application/use-cases/login.use-case';
+import { AuthResponseDto } from '@/domain/dto/auth-reponse.dto';
 
 describe('AuthController - Login', () => {
   let authController: AuthController;
@@ -22,7 +23,16 @@ describe('AuthController - Login', () => {
   });
 
   it('deve realizar login com sucesso e retornar um token', async () => {
-    const mockToken = { access_token: 'jwt-token' };
+    const mockToken: AuthResponseDto = {
+      token: 'token',
+      user: {
+        id: '1',
+        email: '',
+        name: '',
+        role: 'ADMIN',
+        createdAt: new Date(),
+      },
+    };
     jest.spyOn(loginUseCase, 'execute').mockResolvedValue(mockToken);
 
     const result = await authController.login({
@@ -31,7 +41,7 @@ describe('AuthController - Login', () => {
     });
 
     expect(result).toEqual(mockToken);
-    expect(loginUseCase.execute).toHaveBeenCalledWith({
+    expect((loginUseCase.execute as jest.Mock).mock.calls[0][0]).toEqual({
       email: 'test@email.com',
       password: '123456',
     });
