+++ test/controllers/auth/register.test.ts
